Release 2024-01-26

+++ server/pdxmodel/src/main/java/org/msh/pdex2/repository/i18n/ResourceMessageRepo.java
@@ -11,8 +11,8 @@ import org.springframework.data.repository.CrudRepository;
 public interface ResourceMessageRepo extends CrudRepository<ResourceMessage, Long> {
 
 	@Query("SELECT rm FROM ResourceMessage as rm where rm.message_key like :key and key_bundle=:keyBundle")
-	List<ResourceMessage> findByMessage_key(String key, int keyBundle);
+	List<ResourceMessage> findByMessage_key(String key, long keyBundle);
 	
 	@Query("SELECT rm FROM ResourceMessage as rm where key_bundle=:keyBundle")
-	List<ResourceMessage> findByKeyBundle(int keyBundle);
+	List<ResourceMessage> findByKeyBundle(long keyBundle);
 }
